task(UPDATES): Update UAT feedback

+++ src/app/shared/util/timer.service.ts
@@ -3,6 +3,8 @@ import { Router, ActivatedRoute } from '@angular/router'
 import { filter } from "rxjs";
 import { JwtHelperService } from '@auth0/angular-jwt';
 import { AuthService } from "src/app/core/services/auth/auth.service";
+import { Store } from "@ngrx/store";
+import { logoutAction } from "src/app/store/auth/action";
 
 const MINUTES_UNITL_AUTO_LOGOUT = 10 // in mins
 const CHECK_INTERVAL = 5000 // in ms
@@ -26,7 +28,9 @@ export class TimerService {
 
   constructor(private router: Router, 
     private route: ActivatedRoute,
-    private authService: AuthService) {
+    private authService: AuthService,
+    private store: Store,
+    ) {
       const data: any = localStorage.getItem('userData')
       this.userData = JSON.parse(data)
   
@@ -36,24 +40,24 @@ export class TimerService {
     this.initInterval();
     
     localStorage.setItem(STORE_KEY,Date.now().toString());
-    setInterval(() => {
-      if (this.isIdle()) {
-        this.lockUser();
-      }
-    }, 300000 ); // 1 minute in milliseconds
+    // setInterval(() => {
+    //   if (this.isIdle()) {
+    //     this.lockUser();
+    //   }
+    // }, 300000 ); // 1 minute in milliseconds
 }
 
-isIdle() {
-  return !document.body.classList.contains('active');
-}
+// isIdle() {
+//   return !document.body.classList.contains('active');
+// }
 
-lockUser() {
-  const currentURL = window.location.href;
-  const path = new URL(currentURL).pathname;
-  this.router.navigate(['/idle-user'], { queryParams: { returnUrl: path } });
+// lockUser() {
+//   const currentURL = window.location.href;
+//   const path = new URL(currentURL).pathname;
+//   // this.router.navigate(['/idle-user'], { queryParams: { returnUrl: path } });
 
   
-}
+// }
 
   initListener() {
     document.body.addEventListener('click', () => this.reset());
@@ -102,8 +106,13 @@ logOut() {
     emailAddress : this.userData.email
   }
   this.authService.logOut(payload).subscribe((res: any) => {
-    localStorage.clear()
-    this.router.navigateByUrl('/')
+    if (res) {
+
+      localStorage.clear()
+      this.router.navigateByUrl('/')
+      this.store.dispatch(logoutAction());
+
+    }
     })
 }
